backend/requirements.txt
# fastapi==0.110.0
# uvicorn[standard]==0.27.1
# pydantic==2.6.3
# pydantic-settings==2.2.1
# python-multipart==0.0.9
# pyjwt==2.8.0
# sqlmodel==0.0.16
# asyncpg==0.29.0 
# psycopg2-binary==2.9.9  
# psycopg[binary]==3.1.18
# alembic==1.13.1
# celery[redis]==5.3.6
# redis==5.0.1
# boto3==1.34.51
# qdrant-client==1.8.0
# google-generativeai==0.3.2
# langchain==0.1.12
# langchain-google-genai==0.0.9
# langchain-core==0.1.31
# langgraph==0.0.26
# GitPython==3.1.42
# tree-sitter==0.21.0
# tree-sitter-languages==1.10.0
# msgpack==1.0.7
# structlog==24.1.0
# requests==2.31.0
# minio
# # Testing
# pytest==8.0.0
# httpx==0.27.0
# pytest-asyncio==0.23.5


# --- Core Framework ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
pyjwt==2.8.0

# --- Database & ORM ---
sqlmodel==0.0.16
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Task Queue ---
celery[redis]==5.3.6
redis==5.0.1

# --- Storage ---
boto3==1.34.51
minio

# --- Vector & AI (CRITICAL: These versions support structured output) ---
qdrant-client==1.8.0
google-generativeai==0.5.2
langchain==0.1.20
langchain-google-genai==1.0.3
langchain-core==0.1.52
langgraph==0.0.39

# --- Ingestion Engine ---
GitPython==3.1.42
tree-sitter==0.21.0
tree-sitter-languages==1.10.0
msgpack==1.0.7

# --- Observability & Utils ---
structlog==24.1.0
requests==2.31.0
tenacity==8.2.3

# --- Testing ---
pytest==8.0.0
httpx==0.27.0
pytest-asyncio==0.23.5